Add charts for criterias to StudentInfo view.
This also extracts the evaluation informations into an extra component.
